requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.28.0
pydantic>=2.0.0
pyyaml>=6.0
